Log-in Method

`acyrsa_login()` it's used to Log-in and obtained a valid token tthat then should be used in all requests to the API.

acyrsa_login(user, pass, env)

Arguments

user
String. Mandatory User Name

pass
String. Mandatory Password

env
String. Mandatory Wich environment are you going to connect:

demoapi - for testing

api - for production

Value

Creates an 'acyRsaConnection' S4 Object with a token and a base_url.

Note

The token is valid only for a day.

Accesors

You can use accesors to get information about the Object by using:

token(conn)

base_url(conn)

login_date_time(conn)

agent(conn)

user_name(conn)

Examples

if (FALSE) {
conn <- acyrsa_login(user = "asd", pass = "xxx", env = "api")
}